git_tag: Fetch the Makefile for the changed revision
Instead of looking up the "latest" for the branch that
changed, go on and look for the Makefile from the commit that
just changed. This is because, due to delay in the CDN, a
change could be detected but the information on the branch had
yet to be updated; therefore, two lines were printed for the
same commit but with different kernel versions.

Signed-off-by: Daniel Díaz <yosoy@danieldiaz.org>